Fix missing name of stillingskode in kodeverk
Certain Stillingskoder i AAreg are faulty and there will therefore not be returned and Name of Stilling from Kodeverk for these Stillingskoder. This causes exceptions when Arbeidsgiver tries to completed the approval of a Oppfolgingsplan. THis temporary and hacky fix return a Stillingsnavn that is a Stillingskode with a message that describe that the Stillingsode is faulty.